+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `ServiceUser` field `accessControl`, type `object`: AccessControl configures service-specific access control rules for the user.
 When this block is present, the operator manages the full access-control scope it contains
+- Add `OpenSearchACLConfig` to manage OpenSearch ACL
 
 ## v0.36.0 - 2026-03-05